+++ src/entity/leaderboard/LeaderboardByHighScore.java
@@ -1,5 +1,8 @@
-package entity;
+package entity.leaderboard;
 
+
+import entity.user.CommonUser;
+import entity.user.User;
 
 import java.time.LocalTime;
 import java.util.*;
