+++ src/entity/user/CommonUserFactory.java
@@ -1,9 +1,9 @@
-package entity;
+package entity.user;
 
 import java.time.LocalTime;
 import java.util.Map;
 
-public class CommonUserFactory implements UserFactory{
+public class CommonUserFactory implements UserFactory {
 
     @Override
     public User create(String name, String password, Map<LocalTime, Integer> scores){
