+++ src/entity/user/User.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.user;
 
 import entity.board.GameState;
