+++ src/entity/user/UserFactory.java
@@ -1,4 +1,6 @@
-package entity;
+package entity.user;
+
+import entity.user.User;
 
 import java.time.LocalTime;
 import java.util.Map;
